+++ packages/ast/src/clients/rpc/index.ts
@@ -1,3 +1,3 @@
-export * from './class';
+export * from './class/tendermint';
 export * from './extension';
 export * from './scoped';
